Complete function to write spreadsheet 'variables.metadata.xlsx' that contains metadata that can be auto created and provides space for all metadata. This spreadsheet must be edited manually using the questionnaire to assign values of metadata variables for each census variable. Later it will be read back into R and saved as .rds.
This involved writing 'variable.metadata.autocreate.R' and revising 'getvariable.R' and 'getcodebook.R'.

'getvariable.R' was revised to restrict existence check for variable name to mode character--the 'dist' variable was not being read because there is a 'dist' function in the base distribution of R. Seems this fix works, but may have to return to it, I'm unclear on the documentation of exists().